Frontend+Roadmap/backend/requirements.txt
django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
anthropic==0.25.0
python-dotenv==1.0.0
gunicorn==21.2.0
